// src/sdk/database/orm/orm_user_entitlement.ts
export enum UserEntitlementBillingPeriod {
  Unspecified = 0,
  Monthly = 1,
  Yearly = 2,
}
/**
 * Enumeration for UserEntitlementPlanTier
 */
export enum UserEntitlementPlanTier {
  Unspecified = 0,
  Free = 1,
  Pro = 2,
  Premium = 3,
}
// ...
export interface UserEntitlementModel {
  id: string;
  data_creator: string;
  data_updater: string;
  create_time: string;
  update_time: string;
  user_id: string;
  plan_tier: UserEntitlementPlanTier;
  billing_period: UserEntitlementBillingPeriod;
  stripe_session_id?: string | null;
  expires_at?: string | null;
}